+++ pytket/pytket/qasm/includes/_oqclib1_decls.py
@@ -1,0 +1,5 @@
+_INCLUDE_DECLS = {
+    "rz": {"args": ["lam"], "name": "rz"},
+    "sx": {"args": [], "name": "sx"},
+    "ecr": {"args": [], "name": "ecr"},
+}